Swift move DateTests
